src/data_ingestion/requirements.txt
prefect
python-dotenv
pymongo[srv]
